refactor: split engine and hook test domains
