+++ packages/core/src/engine.ts
@@ -1,44 +1,23 @@
 import type {
-  ActiveClip,
-  ActiveClipQuery,
-  ActiveLayerSelector,
-  ActiveLayerOptions,
-  ActiveLayerResult,
   Clip,
-  ClipHitRegion,
-  ClipHitTestInput,
-  ClipHitTestResult,
-  ClipSourceRange,
-  ClipViewportRect,
   TimelineClipDropFeedback,
   Marker,
   PlaybackOptions,
   TimelineClipMoveOptions,
   TimelineClipMoveResult,
-  TimelineClipGeometryOptions,
-  TimelineClipRect,
   TimelineSnapFeedback,
   TimelineSnapResult,
   TimelineSnapTarget,
-  TimelineTrackHeightBatchOptions,
-  TimelineTrackGeometryOptions,
-  TimelineTrackHeightUpdate,
-  TimelineTrackHitTestResult,
-  TimelineTrackRect,
   TimelineEditImpact,
   TimelineEditImpacts,
   TimelineEditPolicy,
   TimelineEditPreview,
-  TimelineInteractionGeometry,
   TimelineClipGroup,
   TimelineKeyframePropertyDefinition,
   TimelineKeyframePropertyId,
   TimelineRegisteredKeyframePropertyDefinition,
   TimelineState,
-  TrackHitTestInput,
   Track,
-  VisibleTimelineClip,
-  VisibleTimelineClipOptions,
 } from '#core/types';
 import type { ClipCreatedEvent, ClipRemovedEvent } from '#core/events';
 import type { RationalTime } from '@techsquidtv/canvas-timeline-utils';
@@ -65,7 +44,6 @@ import {
   createClipSnapshot,
   createClipGroupSnapshots,
   createMarkerSnapshots,
-  createTrackSnapshot,
   createTrackSnapshots,
   stringifyTrackSnapshots,
 } from '#core/snapshot';
@@ -79,11 +57,6 @@ import {
 } from '#core/engine/feedback';
 import {
   defaultTimelineMaxPixelsPerFrame,
-  defaultTimelineViewportHeight,
-  defaultTimelineViewportWidth,
-  normalizeViewportCoordinate,
-  resolveTimelineInteractionGeometry,
-  type ResolvedTimelineInteractionGeometry,
   type TimelineZoomConstraints,
 } from '#core/engine/geometry';
 import { filterClipKeyframesToClipRange, shiftClipKeyframes } from '#core/engine/clip-keyframes';
@@ -92,13 +65,7 @@ import type {
   TimelineSnapProvider,
   TimelineSnapProviderContext,
 } from '#core/engine/snapping';
-import {
-  createClipSourceRange,
-  createClipSyncKey,
-  mapSourceTimeToTimelineTime,
-  mapTimelineTimeToSourceTime,
-} from '#core/engine/media-sync';
-import { TimelineEngineKeyframes } from '#core/engine/keyframes';
+import { TimelineEngineState } from '#core/engine/timeline-state';
 
 export {
   defaultTimelineInteractionGeometry,
@@ -121,7 +88,7 @@ export type {
  * builds dynamic snap indexes for magnetic snap guidance, handles split and edit actions,
  * and publishes state events to trigger low-latency canvas renderings and lightweight React layouts.
  */
-export class TimelineEngine extends TimelineEngineKeyframes {
+export class TimelineEngine extends TimelineEngineState {
   /**
    * Creates an instance of the TimelineEngine.
    *
@@ -213,7 +180,7 @@ export class TimelineEngine extends TimelineEngineKeyframes {
   /**
    * Monotonic revision for changes that can affect active layer lookup.
    */
-  get contentRevision() {
+  override get contentRevision() {
     return this.state.contentRevision;
   }
 
@@ -675,7 +642,7 @@ export class TimelineEngine extends TimelineEngineKeyframes {
 
   // --- Helpers ---
 
-  private hasZoomConstraints() {
+  protected override hasZoomConstraints() {
     return (
       this.zoomConstraints.frameRate !== undefined ||
       this.zoomConstraints.maxPixelsPerFrame !== undefined ||
@@ -730,717 +697,8 @@ export class TimelineEngine extends TimelineEngineKeyframes {
     return maxZoomScale;
   }
 
-  private clampZoomScale(scale: number) {
+  protected override clampZoomScale(scale: number) {
     return Math.max(this.minZoomScale, Math.min(scale, this.maxZoomScale));
-  }
-
-  /**
-   * Locates a clip and returns its containing track and indexes.
-   *
-   * @param clipId - Clip id to look up.
-   * @returns Clip lookup details, or `undefined` when the clip is not found.
-   */
-  getClip(
-    clipId: string
-  ): { track: Track; clip: Clip; trackIndex: number; clipIndex: number } | undefined {
-    for (let trackIndex = 0; trackIndex < this.state.tracks.length; trackIndex++) {
-      const track = this.state.tracks[trackIndex];
-      for (let clipIndex = 0; clipIndex < track.clips.length; clipIndex++) {
-        const clip = track.clips[clipIndex];
-        if (clip.id === clipId) {
-          return { track, clip, trackIndex, clipIndex };
-        }
-      }
-    }
-    return undefined;
-  }
-
-  /**
-   * Returns viewport rectangles for every track row in track order.
-   *
-   * @param options - Optional ruler, track metrics, and viewport width.
-   * @returns Track row entries matching canvas layout.
-   */
-  getTrackRects(options: TimelineTrackGeometryOptions = {}): TimelineTrackRect[] {
-    const resolvedGeometry = resolveTimelineInteractionGeometry(options);
-    const viewportWidth = Math.max(
-      0,
-      options.viewportWidth ?? this.state.viewportWidth ?? defaultTimelineViewportWidth
-    );
-    const scrollTop = this.state.scrollTop;
-    const trackRects: TimelineTrackRect[] = [];
-    let y = resolvedGeometry.rulerHeight - scrollTop;
-
-    for (let trackIndex = 0; trackIndex < this.state.tracks.length; trackIndex++) {
-      const track = this.state.tracks[trackIndex];
-      const height = this.getTrackViewportHeight(track, resolvedGeometry);
-      trackRects.push(this.createTrackViewportRect(track, trackIndex, y, height, viewportWidth));
-      y += height;
-    }
-
-    return trackRects;
-  }
-
-  /**
-   * Hit-tests timeline tracks in viewport coordinates, matching canvas layout.
-   *
-   * @param input - Viewport point and optional geometry overrides.
-   * @returns The matching track row, or null for blank/ruler space.
-   */
-  getTrackAtPoint<TrackKind = string>(
-    input: TrackHitTestInput
-  ): TimelineTrackHitTestResult<TrackKind> | null {
-    const resolvedGeometry = resolveTimelineInteractionGeometry(input);
-    if (input.y < resolvedGeometry.rulerHeight) {
-      return null;
-    }
-
-    for (const rect of this.getTrackRects(input)) {
-      const track = this.getTracks<TrackKind>()[rect.trackIndex];
-      if (track === undefined) {
-        continue;
-      }
-      const insideX =
-        input.x === undefined || (input.x >= rect.x && input.x <= rect.x + rect.width);
-
-      if (insideX && input.y >= rect.y && input.y < rect.y + rect.height) {
-        return { track, trackIndex: rect.trackIndex, rect };
-      }
-    }
-
-    return null;
-  }
-
-  /**
-   * Returns the current viewport rectangle for a clip, matching canvas track layout.
-   *
-   * @param clipId - Clip id to locate.
-   * @param geometry - Optional ruler and track metrics used by the interaction layer.
-   * @returns Viewport rectangle for the clip, or `null` when the clip is missing.
-   */
-  getClipRect(clipId: string, geometry: TimelineInteractionGeometry = {}): ClipViewportRect | null {
-    const resolvedGeometry = resolveTimelineInteractionGeometry(geometry);
-    let y = resolvedGeometry.rulerHeight - this.state.scrollTop;
-
-    for (let trackIndex = 0; trackIndex < this.state.tracks.length; trackIndex++) {
-      const track = this.state.tracks[trackIndex];
-      const trackHeight = this.getTrackViewportHeight(track, resolvedGeometry);
-
-      for (let clipIndex = 0; clipIndex < track.clips.length; clipIndex++) {
-        const clip = track.clips[clipIndex];
-        if (clip.id !== clipId) {
-          continue;
-        }
-
-        return this.createClipViewportRect(track, clip, trackIndex, clipIndex, y, trackHeight);
-      }
-
-      y += trackHeight;
-    }
-
-    return null;
-  }
-
-  /**
-   * Returns viewport rectangles for every clip in track order.
-   *
-   * This is the canonical geometry source for DOM clip rendering and custom
-   * canvas layers. Pointer hit testing keeps a direct low-allocation path.
-   *
-   * @param options - Optional ruler and track metrics used to align with the renderer.
-   * @returns Clip entries with viewport rectangles and edit/display state.
-   */
-  getClipRects<TrackKind = string>(
-    options: TimelineClipGeometryOptions = {}
-  ): TimelineClipRect<TrackKind>[] {
-    const clipRects: TimelineClipRect<TrackKind>[] = [];
-
-    this.forEachTimelineClipGeometry<TrackKind>(
-      options,
-      (track, clip, trackIndex, clipIndex, rect) => {
-        clipRects.push(this.createTimelineClipRect(track, clip, trackIndex, clipIndex, rect));
-      }
-    );
-
-    return clipRects;
-  }
-
-  /**
-   * Returns clips intersecting the current horizontal viewport, plus optional overscan.
-   *
-   * The returned visible ranges map clipped viewport pixels back to timeline and
-   * source-media times so custom renderers can draw cached thumbnails, waveforms,
-   * annotations, or heatmaps without duplicating timeline math.
-   *
-   * @param options - Viewport, overscan, and optional geometry settings.
-   * @returns Viewport-intersecting clip entries in track order.
-   */
-  getVisibleTimelineClips<TrackKind = string>(
-    options: VisibleTimelineClipOptions = {}
-  ): VisibleTimelineClip<TrackKind>[] {
-    const resolvedGeometry = resolveTimelineInteractionGeometry(options);
-    const viewportWidth = Math.max(
-      0,
-      options.viewportWidth ?? this.state.viewportWidth ?? defaultTimelineViewportWidth
-    );
-    const viewportHeight =
-      options.viewportHeight === undefined ? undefined : Math.max(0, options.viewportHeight);
-    const overscanPixels = Math.max(0, options.overscanPixels ?? 0);
-    const minX = -overscanPixels;
-    const maxX = viewportWidth + overscanPixels;
-    const minY = resolvedGeometry.rulerHeight - overscanPixels;
-    const maxY = viewportHeight === undefined ? undefined : viewportHeight + overscanPixels;
-    const visibleClips: VisibleTimelineClip<TrackKind>[] = [];
-
-    this.forEachTimelineClipGeometry<TrackKind>(
-      options,
-      (track, clip, trackIndex, clipIndex, rect) => {
-        const rectRight = rect.x + rect.width;
-        const rectBottom = rect.y + rect.height;
-
-        if (rectRight < minX || rect.x > maxX) {
-          return;
-        }
-        if (maxY !== undefined && (rectBottom < minY || rect.y > maxY)) {
-          return;
-        }
-
-        const visibleLeft = Math.max(rect.x, minX);
-        const visibleRight = Math.min(rectRight, maxX);
-        const visibleTop = maxY === undefined ? rect.y : Math.max(rect.y, minY);
-        const visibleBottom = maxY === undefined ? rectBottom : Math.min(rectBottom, maxY);
-        const visibleWidth = Math.max(0, visibleRight - visibleLeft);
-        const visibleHeight = Math.max(0, visibleBottom - visibleTop);
-        const visibleTimelineStartTime = maxRational(
-          clip.timelineStart,
-          this.pixelToTime(visibleLeft, clip.timelineStart.r)
-        );
-        const visibleTimelineEndTime = minRational(
-          clip.timelineEnd,
-          this.pixelToTime(visibleRight, clip.timelineEnd.r)
-        );
-
-        if (compareRational(visibleTimelineEndTime, visibleTimelineStartTime) <= 0) {
-          return;
-        }
-
-        const visibleSourceStartTime = addRational(
-          clip.sourceStart,
-          subRational(visibleTimelineStartTime, clip.timelineStart)
-        );
-        const visibleSourceEndTime = addRational(
-          clip.sourceStart,
-          subRational(visibleTimelineEndTime, clip.timelineStart)
-        );
-        const clipRect = this.createTimelineClipRect(track, clip, trackIndex, clipIndex, rect);
-
-        visibleClips.push({
-          ...clipRect,
-          visibleRect: {
-            clipId: clip.id,
-            trackId: track.id,
-            trackIndex,
-            clipIndex,
-            x: normalizeViewportCoordinate(visibleLeft),
-            y: normalizeViewportCoordinate(visibleTop),
-            width: normalizeViewportCoordinate(visibleWidth),
-            height: normalizeViewportCoordinate(visibleHeight),
-          },
-          visibleTimelineStartTime,
-          visibleTimelineEndTime,
-          visibleSourceStartTime,
-          visibleSourceEndTime,
-        });
-      }
-    );
-
-    return visibleClips;
-  }
-
-  /**
-   * Hit-tests timeline clips in viewport coordinates, matching canvas track layout.
-   *
-   * Clip bodies remain selectable even when their track is locked or the clip is
-   * non-editable; edge/body edit regions are only reported when the clip can be edited.
-   *
-   * @param input - Viewport point and optional geometry overrides.
-   * @returns The topmost matching clip hit, or `null` for blank/ruler space.
-   */
-  getClipAtPoint(input: ClipHitTestInput): ClipHitTestResult | null {
-    const resolvedGeometry = resolveTimelineInteractionGeometry(input);
-    if (input.y < resolvedGeometry.rulerHeight) {
-      return null;
-    }
-
-    let trackY = resolvedGeometry.rulerHeight - this.state.scrollTop;
-
-    for (let trackIndex = 0; trackIndex < this.state.tracks.length; trackIndex++) {
-      const track = this.state.tracks[trackIndex];
-      const trackHeight = this.getTrackViewportHeight(track, resolvedGeometry);
-      const trackBottom = trackY + trackHeight;
-
-      if (input.y >= trackY && input.y < trackBottom) {
-        return (
-          this.findClipHitInTrack(track, trackIndex, trackY, trackHeight, resolvedGeometry, input, {
-            selectedOnly: true,
-          }) ??
-          this.findClipHitInTrack(track, trackIndex, trackY, trackHeight, resolvedGeometry, input, {
-            selectedOnly: false,
-            skipSelected: true,
-          })
-        );
-      }
-
-      trackY = trackBottom;
-    }
-
-    return null;
-  }
-
-  private getTrackViewportHeight<TrackKind>(
-    track: Track<TrackKind>,
-    geometry: ResolvedTimelineInteractionGeometry
-  ): number {
-    return Math.floor(
-      track.collapsed ? geometry.collapsedTrackHeight : (track.height ?? geometry.trackHeight)
-    );
-  }
-
-  protected override forEachTimelineClipGeometry<TrackKind>(
-    options: TimelineClipGeometryOptions,
-    visit: (
-      track: Track<TrackKind>,
-      clip: Clip,
-      trackIndex: number,
-      clipIndex: number,
-      rect: ClipViewportRect
-    ) => void
-  ): void {
-    const resolvedGeometry = resolveTimelineInteractionGeometry(options);
-    let y = resolvedGeometry.rulerHeight - this.state.scrollTop;
-
-    for (let trackIndex = 0; trackIndex < this.state.tracks.length; trackIndex++) {
-      const track = this.getTracks<TrackKind>()[trackIndex];
-      const trackHeight = this.getTrackViewportHeight(track, resolvedGeometry);
-
-      for (let clipIndex = 0; clipIndex < track.clips.length; clipIndex++) {
-        const clip = track.clips[clipIndex];
-        const rect = this.createClipViewportRect(
-          track,
-          clip,
-          trackIndex,
-          clipIndex,
-          y,
-          trackHeight
-        );
-
-        visit(track, clip, trackIndex, clipIndex, rect);
-      }
-
-      y += trackHeight;
-    }
-  }
-
-  private findClipHitInTrack(
-    track: Track,
-    trackIndex: number,
-    trackY: number,
-    trackHeight: number,
-    geometry: ResolvedTimelineInteractionGeometry,
-    input: ClipHitTestInput,
-    options: { selectedOnly?: boolean; skipSelected?: boolean } = {}
-  ): ClipHitTestResult | null {
-    for (let clipIndex = track.clips.length - 1; clipIndex >= 0; clipIndex--) {
-      const clip = track.clips[clipIndex];
-      if (options.selectedOnly && !clip.selected) {
-        continue;
-      }
-      if (options.skipSelected && clip.selected) {
-        continue;
-      }
-
-      const rect = this.createClipViewportRect(
-        track,
-        clip,
-        trackIndex,
-        clipIndex,
-        trackY,
-        trackHeight
-      );
-
-      if (input.x < rect.x || input.x > rect.x + rect.width) {
-        continue;
-      }
-
-      const canMove = !track.locked && clip.movable !== false;
-      const canTrim = !track.locked && clip.resizable !== false;
-      let region: ClipHitRegion = 'body';
-
-      if (canTrim) {
-        const baseThreshold =
-          input.pointerType === 'touch' ? geometry.touchEdgeThreshold : geometry.edgeThreshold;
-        const edgeThreshold = Math.min(baseThreshold, rect.width / 3);
-
-        if (input.x - rect.x < edgeThreshold) {
-          region = 'start-edge';
-        } else if (rect.x + rect.width - input.x < edgeThreshold) {
-          region = 'end-edge';
-        }
-      }
-
-      return {
-        track,
-        clip,
-        trackIndex,
-        clipIndex,
-        region,
-        rect,
-        canMove,
-        canTrim,
-      };
-    }
-
-    return null;
-  }
-
-  private createClipViewportRect<TrackKind>(
-    track: Track<TrackKind>,
-    clip: Clip,
-    trackIndex: number,
-    clipIndex: number,
-    y: number,
-    height: number
-  ): ClipViewportRect {
-    const x = this.timeToPixel(clip.timelineStart);
-    const endX = this.timeToPixel(clip.timelineEnd);
-
-    return {
-      clipId: clip.id,
-      trackId: track.id,
-      trackIndex,
-      clipIndex,
-      x,
-      y,
-      width: endX - x,
-      height,
-    };
-  }
-
-  private createTrackViewportRect<TrackKind>(
-    track: Track<TrackKind>,
-    trackIndex: number,
-    y: number,
-    height: number,
-    width: number
-  ): TimelineTrackRect {
-    return {
-      trackId: track.id,
-      trackIndex,
-      x: 0,
-      y,
-      width,
-      height,
-    };
-  }
-
-  private createTimelineClipRect<TrackKind>(
-    track: Track<TrackKind>,
-    clip: Clip,
-    trackIndex: number,
-    clipIndex: number,
-    rect: ClipViewportRect
-  ): TimelineClipRect<TrackKind> {
-    return {
-      clip,
-      track,
-      trackIndex,
-      clipIndex,
-      rect,
-      sourceRange: createClipSourceRange(clip),
-      canMove: !track.locked && clip.movable !== false,
-      canTrim: !track.locked && clip.resizable !== false,
-      muted: track.muted,
-      visible: track.visible,
-      locked: track.locked,
-      disabled: clip.disabled === true,
-    };
-  }
-
-  /**
-   * Marks timeline content as changed and notifies subscribers.
-   *
-   * Use this after external metadata that affects rendering changes without a
-   * structural timeline edit, such as waveform availability, thumbnails, or
-   * cached analysis keyed by clip id.
-   */
-  invalidateContent() {
-    this.state.contentRevision = this.contentRevision + 1;
-    this.emit('content:change', this.state.contentRevision);
-  }
-
-  /**
-   * Returns enabled clips under a timeline time, ordered by track and clip order.
-   *
-   * Hidden or muted tracks and disabled clips are excluded so preview and playback
-   * integrations can pick sources using the same visibility rules. Clip
-   * intervals are half-open: active at `timelineStart` and inactive at
-   * `timelineEnd`.
-   *
-   * @param time - Timeline time to inspect. Defaults to the current playhead.
-   * @returns Active clips with computed source-media timestamps.
-   */
-  getActiveClips(time: RationalTime = this.state.playheadTime): ActiveClip[] {
-    const activeClips: ActiveClip[] = [];
-
-    for (const track of this.state.tracks) {
-      if (!track.visible || track.muted) {
-        continue;
-      }
-      for (const clip of track.clips) {
-        if (clip.disabled) {
-          continue;
-        }
-        const activeClip = this.createActiveClip(track, clip, time);
-        if (activeClip !== undefined) {
-          activeClips.push(activeClip);
-        }
-      }
-    }
-
-    return activeClips;
-  }
-
-  /**
-   * Returns the first active clip matching an active layer query.
-   *
-   * @param query - Optional timeline time, track kind, source id, and custom predicate.
-   * @returns First matching active clip in track and clip order.
-   */
-  getActiveClip(query: ActiveClipQuery = {}): ActiveClip | undefined {
-    const { time = this.state.playheadTime, trackKind, sourceId, predicate } = query;
-    return this.getActiveClips(time).find((activeClip) => {
-      return this.matchesActiveLayerSelector(activeClip, { trackKind, sourceId, predicate });
-    });
-  }
-
-  /**
-   * Finds the clips active at a timeline time and groups them by named layer selectors.
-   *
-   * Use this for preview, playback, subtitle, and effect integrations that need
-   * to know which layer clips should be rendered or synced at the playhead. Hidden
-   * or muted tracks and disabled clips are excluded. The result preserves every match in
-   * `layers`, deduplicates matches in `all`, and exposes `primary` as a
-   * first-match convenience for each requested layer.
-   *
-   * @param options - Timeline time and named active layer selectors.
-   * @returns Active layer result for the requested layers.
-   */
-  getActiveLayers<LayerName extends string = string>(
-    options: ActiveLayerOptions<LayerName>
-  ): ActiveLayerResult<LayerName> {
-    const time = options.time ?? this.state.playheadTime;
-    const activeClips = this.getActiveClips(time);
-    const layers = {} as Record<LayerName, ActiveClip[]>;
-    const primary: Partial<Record<LayerName, ActiveClip>> = {};
-    const matchedClips = new Map<string, ActiveClip>();
-
-    for (const layerName of Object.keys(options.layers) as LayerName[]) {
-      const selector = options.layers[layerName];
-      const layerClips = activeClips.filter((activeClip) =>
-        this.matchesActiveLayerSelector(activeClip, selector)
-      );
-      layers[layerName] = layerClips;
-      if (layerClips[0] !== undefined) {
-        primary[layerName] = layerClips[0];
-      }
-      for (const activeClip of layerClips) {
-        if (!matchedClips.has(activeClip.clip.id)) {
-          matchedClips.set(activeClip.clip.id, activeClip);
-        }
-      }
-    }
-
-    const all = [...matchedClips.values()];
-    const byTrack = this.groupActiveClipsByTrack(all);
-
-    return {
-      time,
-      all,
-      byTrack,
-      layers,
-      primary,
-      hasActiveClips: all.length > 0,
-      firstContentTime: this.getFirstContentTime({ layers: options.layers }),
-    };
-  }
-
-  /**
-   * Finds the earliest clip start matching any requested layer.
-   *
-   * @param options - Named layer selectors.
-   * @returns Earliest matching timeline start, or `undefined` when nothing matches.
-   */
-  getFirstContentTime<LayerName extends string = string>(
-    options: Pick<ActiveLayerOptions<LayerName>, 'layers'>
-  ): RationalTime | undefined {
-    let firstContentTime: RationalTime | undefined;
-
-    for (const track of this.state.tracks) {
-      if (!track.visible || track.muted) {
-        continue;
-      }
-
-      for (const clip of track.clips) {
-        if (clip.disabled) {
-          continue;
-        }
-
-        const activeClip = this.createActiveClip(track, clip, clip.timelineStart);
-        if (
-          activeClip === undefined ||
-          !Object.values<ActiveLayerSelector>(options.layers).some((selector) =>
-            this.matchesActiveLayerSelector(activeClip, selector)
-          )
-        ) {
-          continue;
-        }
-
-        if (
-          firstContentTime === undefined ||
-          compareRational(clip.timelineStart, firstContentTime) < 0
-        ) {
-          firstContentTime = clip.timelineStart;
-        }
-      }
-    }
-
-    return firstContentTime;
-  }
-
-  /**
-   * Groups active clips by containing track id.
-   *
-   * @param time - Timeline time to inspect. Defaults to the current playhead.
-   * @returns Map of track id to active clips on that track.
-   */
-  getActiveClipsByTrack(time: RationalTime = this.state.playheadTime): Map<string, ActiveClip[]> {
-    return this.groupActiveClipsByTrack(this.getActiveClips(time));
-  }
-
-  private groupActiveClipsByTrack(activeClips: ActiveClip[]): Map<string, ActiveClip[]> {
-    const clipsByTrack = new Map<string, ActiveClip[]>();
-    for (const activeClip of activeClips) {
-      const trackClips = clipsByTrack.get(activeClip.track.id);
-      if (trackClips === undefined) {
-        clipsByTrack.set(activeClip.track.id, [activeClip]);
-      } else {
-        trackClips.push(activeClip);
-      }
-    }
-    return clipsByTrack;
-  }
-
-  private matchesActiveLayerSelector(
-    activeClip: ActiveClip,
-    selector: ActiveLayerSelector
-  ): boolean {
-    if (selector.trackKind !== undefined && activeClip.track.kind !== selector.trackKind) {
-      return false;
-    }
-    if (selector.sourceId !== undefined && activeClip.clip.sourceId !== selector.sourceId) {
-      return false;
-    }
-    return selector.predicate === undefined || selector.predicate(activeClip);
-  }
-
-  private createActiveClip(
-    track: Track,
-    clip: Clip,
-    timelineTime: RationalTime
-  ): ActiveClip | undefined {
-    const sourceTime = this.timelineTimeToSourceTime(clip, timelineTime);
-    const sourceRange = this.getClipSourceRange(clip);
-    const syncKey = this.getClipSyncKey(clip);
-    if (sourceTime === undefined || sourceRange === undefined || syncKey === undefined) {
-      return undefined;
-    }
-
-    return {
-      track,
-      clip,
-      timelineTime,
-      sourceTime,
-      sourceRange,
-      syncKey,
-    };
-  }
-
-  /**
-   * Computes the source-media range covered by a clip.
-   *
-   * @param clipIdOrClip - Clip object or id to inspect.
-   * @returns Source range, or `undefined` when the clip is missing.
-   */
-  getClipSourceRange(clipIdOrClip: string | Clip): ClipSourceRange | undefined {
-    const clip = this.resolveClip(clipIdOrClip);
-    if (clip === undefined) {
-      return undefined;
-    }
-
-    return createClipSourceRange(clip);
-  }
-
-  /**
-   * Returns a stable media sync signature for timing-affecting clip fields.
-   *
-   * @param clipIdOrClip - Clip object or id to inspect.
-   * @returns Sync key, or `undefined` when the clip is missing.
-   */
-  getClipSyncKey(clipIdOrClip: string | Clip): string | undefined {
-    const clip = this.resolveClip(clipIdOrClip);
-    if (clip === undefined) {
-      return undefined;
-    }
-
-    return createClipSyncKey(clip);
-  }
-
-  /**
-   * Maps a timeline timestamp within a clip to the matching source-media time.
-   *
-   * @param clipIdOrClip - Clip object or id to map through.
-   * @param timelineTime - Timeline time to convert. Defaults to the current playhead.
-   * @returns Source time, or `undefined` when the clip is missing or the time is outside the clip.
-   */
-  timelineTimeToSourceTime(
-    clipIdOrClip: string | Clip,
-    timelineTime: RationalTime = this.state.playheadTime
-  ): RationalTime | undefined {
-    const clip = this.resolveClip(clipIdOrClip);
-    if (clip === undefined) {
-      return undefined;
-    }
-
-    return mapTimelineTimeToSourceTime(clip, timelineTime);
-  }
-
-  /**
-   * Maps a source-media timestamp within a clip back to timeline time.
-   *
-   * @param clipIdOrClip - Clip object or id to map through.
-   * @param sourceTime - Source-media time to convert.
-   * @returns Timeline time, or `undefined` when the clip is missing or the source time is outside the clip.
-   */
-  sourceTimeToTimelineTime(
-    clipIdOrClip: string | Clip,
-    sourceTime: RationalTime
-  ): RationalTime | undefined {
-    const clip = this.resolveClip(clipIdOrClip);
-    if (clip === undefined) {
-      return undefined;
-    }
-
-    return mapSourceTimeToTimelineTime(clip, sourceTime);
   }
 
   /**
@@ -1620,7 +878,7 @@ export class TimelineEngine extends TimelineEngineKeyframes {
    * @param rate - Tick rate for the returned rational time.
    * @returns Timeline time represented by the pixel.
    */
-  pixelToTime(pixel: number, rate: number = 24000): RationalTime {
+  override pixelToTime(pixel: number, rate: number = 24000): RationalTime {
     assertValidTimelineNumber(pixel, 'pixel');
     return fromSeconds((pixel + this.state.scrollLeft) / this.state.zoomScale, rate);
   }
@@ -1641,7 +899,7 @@ export class TimelineEngine extends TimelineEngineKeyframes {
    *
    * @param time - Desired playhead time.
    */
-  updatePlayhead(time: RationalTime) {
+  override updatePlayhead(time: RationalTime) {
     assertValidRationalTime(time, 'time');
     let clampedTime = maxRational({ v: 0, r: time.r }, time);
     clampedTime = minRational(clampedTime, this.maxContentTime);
@@ -2295,425 +1553,5 @@ export class TimelineEngine extends TimelineEngineKeyframes {
       return true;
     }
     return false;
-  }
-
-  /**
-   * Current marker list.
-   */
-  get markers() {
-    return this.state.markers ?? [];
-  }
-
-  /**
-   * Adds a marker at a timeline time.
-   *
-   * @param time - Timeline time for the marker.
-   * @param label - Optional visible marker label.
-   * @param color - Optional marker color.
-   * @param description - Optional longer marker note.
-   * @returns The created marker.
-   */
-  addMarker(time: RationalTime, label?: string, color?: string, description?: string) {
-    assertValidRationalTime(time, 'time');
-    const marker = {
-      id: crypto.randomUUID(),
-      time,
-      label,
-      color,
-      description,
-    };
-    if (!this.state.markers) {
-      this.state.markers = [];
-    }
-    this.state.markers.push(marker);
-    this.snapshot();
-    this.emit('marker:add', { marker });
-    this.emit('state:settled');
-    this.emit('render');
-    return marker;
-  }
-
-  /**
-   * Removes a marker by id.
-   *
-   * @param id - Marker id to remove.
-   * @returns Whether the marker was found and removed.
-   */
-  removeMarker(id: string) {
-    if (this.state.markers) {
-      const idx = this.state.markers.findIndex((m) => m.id === id);
-      if (idx !== -1) {
-        const removed = this.state.markers.splice(idx, 1)[0];
-        this.snapshot();
-        this.emit('marker:remove', { marker: removed });
-        this.emit('state:settled');
-        this.emit('render');
-        return true;
-      }
-    }
-    return false;
-  }
-
-  /**
-   * Updates an existing marker.
-   *
-   * @param id - Marker id to update.
-   * @param updates - Marker fields to merge into the existing marker.
-   * @returns The updated marker, or `null` when no marker was found.
-   */
-  updateMarker(id: string, updates: Partial<Marker>) {
-    if (updates.time !== undefined) {
-      assertValidRationalTime(updates.time, 'updates.time');
-    }
-    if (this.state.markers) {
-      const marker = this.state.markers.find((m) => m.id === id);
-      if (marker) {
-        Object.assign(marker, updates);
-        this.snapshot();
-        this.emit('marker:update', { marker });
-        this.emit('state:settled');
-        this.emit('render');
-        return marker;
-      }
-    }
-    return null;
-  }
-
-  /**
-   * Appends a track to the timeline.
-   *
-   * @param track - Track to add. Its id should be unique within the timeline.
-   */
-  addTrack(track: Track) {
-    const nextTrack = createTrackSnapshot(track);
-    this.state.tracks.push(nextTrack);
-    this.snapshot();
-    this.emit('track:add', { track: nextTrack });
-    this.invalidateContent();
-    this.emit('state:settled');
-    this.emit('render');
-  }
-
-  /**
-   * Removes a track by id.
-   *
-   * @param trackId - Track id to remove.
-   * @returns Whether the track was found and removed.
-   */
-  removeTrack(trackId: string) {
-    const idx = this.state.tracks.findIndex((t) => t.id === trackId);
-    if (idx !== -1) {
-      const removed = this.state.tracks.splice(idx, 1)[0];
-      const scrollChanged = this.clampScrollTop();
-      this.normalizeClipGroups();
-      this.snapshot();
-      this.emit('track:remove', { track: removed });
-      this.invalidateContent();
-      if (scrollChanged) {
-        this.emitScrollChange();
-      }
-      this.emit('state:settled');
-      this.emit('render');
-      return true;
-    }
-    return false;
-  }
-
-  /**
-   * Enables, disables, or toggles a track's muted state.
-   *
-   * @param trackId - Track id to update.
-   * @param muted - Explicit muted state, or omitted to toggle.
-   */
-  toggleMuteTrack(trackId: string, muted?: boolean) {
-    const track = this.state.tracks.find((t) => t.id === trackId);
-    if (track) {
-      track.muted = muted !== undefined ? muted : !track.muted;
-      this.snapshot();
-      this.emit('track:mute', { trackId: track.id, muted: track.muted });
-      this.invalidateContent();
-      this.emit('state:settled');
-      this.emit('render');
-    }
-  }
-
-  /**
-   * Enables, disables, or toggles a track's output visibility.
-   *
-   * @param trackId - Track id to update.
-   * @param visible - Explicit visible state, or omitted to toggle.
-   */
-  toggleTrackVisibility(trackId: string, visible?: boolean) {
-    const track = this.state.tracks.find((t) => t.id === trackId);
-    if (track) {
-      track.visible = visible !== undefined ? visible : !track.visible;
-      this.snapshot();
-      this.emit('track:visibility', { trackId: track.id, visible: track.visible });
-      this.invalidateContent();
-      this.emit('state:settled');
-      this.emit('render');
-    }
-  }
-
-  /**
-   * Enables, disables, or toggles a track's locked state.
-   *
-   * @param trackId - Track id to update.
-   * @param locked - Explicit locked state, or omitted to toggle.
-   */
-  toggleLockTrack(trackId: string, locked?: boolean) {
-    const track = this.state.tracks.find((t) => t.id === trackId);
-    if (track) {
-      track.locked = locked !== undefined ? locked : !track.locked;
-      this.emit('track:lock', { trackId: track.id, locked: track.locked });
-      this.emit('state:settled');
-      this.emit('render');
-    }
-  }
-
-  /**
-   * Selects one track and clears selection from all others.
-   *
-   * @param trackId - Track id to select, or `null` to clear track selection.
-   */
-  selectTrack(trackId: string | null) {
-    for (const track of this.state.tracks) {
-      track.selected = track.id === trackId;
-    }
-    this.emit('track:select', { trackId });
-    this.emit('state:settled');
-    this.emit('render');
-  }
-
-  /**
-   * Updates a track's expanded display height.
-   *
-   * @param trackId - Track id to resize.
-   * @param height - Expanded row height in pixels.
-   */
-  setTrackHeight(trackId: string, height: number) {
-    assertPositiveTimelineNumber(height, 'height');
-    this.setTrackHeights([{ trackId, height }]);
-  }
-
-  /**
-   * Sets multiple expanded track heights and publishes a single settled/render cycle.
-   *
-   * @param updates - Track height updates to apply.
-   * @param options - Optional viewport state to batch with the height changes.
-   */
-  setTrackHeights(
-    updates: readonly TimelineTrackHeightUpdate[],
-    options: TimelineTrackHeightBatchOptions = {}
-  ) {
-    const resizeEvents: TimelineTrackHeightUpdate[] = [];
-    const previousScrollTop = this.state.scrollTop;
-
-    for (const update of updates) {
-      assertPositiveTimelineNumber(update.height, `height for track "${update.trackId}"`);
-      const track = this.state.tracks.find((t) => t.id === update.trackId);
-      if (!track || track.height === update.height) {
-        continue;
-      }
-
-      track.height = update.height;
-      resizeEvents.push({ trackId: track.id, height: update.height });
-    }
-
-    if (options.scrollTop !== undefined) {
-      assertNonNegativeTimelineNumber(options.scrollTop, 'options.scrollTop');
-      this.state.scrollTop = options.scrollTop;
-    }
-
-    this.state.scrollTop = Math.max(0, Math.min(this.state.scrollTop, this.maxScrollTop));
-    const scrollChanged = this.state.scrollTop !== previousScrollTop;
-
-    if (resizeEvents.length === 0 && !scrollChanged) {
-      return;
-    }
-
-    for (const resizeEvent of resizeEvents) {
-      this.emit('track:resize', resizeEvent);
-    }
-    if (scrollChanged) {
-      this.emitScrollChange();
-    }
-    this.emit('state:settled');
-    this.emit('render');
-  }
-
-  /**
-   * Maximum timeline time used for playback and scroll clamping.
-   */
-  get maxContentTime(): RationalTime {
-    if (this.state.duration !== undefined) {
-      return this.state.duration;
-    }
-    let max = { v: 0, r: 24000 };
-    for (const track of this.state.tracks) {
-      if (track.clips) {
-        for (const clip of track.clips) {
-          if (compareRational(clip.timelineEnd, max) > 0) {
-            max = clip.timelineEnd;
-          }
-        }
-      }
-    }
-    return max;
-  }
-
-  /**
-   * Maximum horizontal scroll offset for the current viewport and content duration.
-   */
-  get maxScrollLeft() {
-    const viewportWidth = this.state.viewportWidth || 1000;
-    const contentEndX = toSeconds(this.maxContentTime) * this.state.zoomScale;
-    return Math.max(0, contentEndX - viewportWidth);
-  }
-
-  /**
-   * Maximum vertical scroll offset for the current viewport and track stack.
-   */
-  get maxScrollTop() {
-    const viewportHeight = this.state.viewportHeight ?? defaultTimelineViewportHeight;
-    const contentHeight = this.getContentHeight();
-    return Math.max(0, contentHeight - viewportHeight);
-  }
-
-  private getContentHeight(geometry: TimelineInteractionGeometry = {}) {
-    const resolvedGeometry = resolveTimelineInteractionGeometry(geometry);
-    return this.state.tracks.reduce(
-      (height, track) => height + this.getTrackViewportHeight(track, resolvedGeometry),
-      resolvedGeometry.rulerHeight
-    );
-  }
-
-  private emitScrollChange() {
-    this.emit('scroll:change', {
-      scrollLeft: this.state.scrollLeft,
-      scrollTop: this.state.scrollTop,
-    });
-  }
-
-  private emitViewportResize() {
-    this.emit('viewport:resize', {
-      viewportWidth: this.state.viewportWidth,
-      viewportHeight: this.state.viewportHeight,
-    });
-  }
-
-  private clampScrollTop() {
-    const clampedScrollTop = Math.max(0, Math.min(this.state.scrollTop, this.maxScrollTop));
-    const changed = clampedScrollTop !== this.state.scrollTop;
-    this.state.scrollTop = clampedScrollTop;
-    return changed;
-  }
-
-  /**
-   * Sets the zoom scale while keeping the viewport within content bounds.
-   *
-   * @param scale - Desired pixels-per-second zoom scale.
-   */
-  setZoomScale(scale: number) {
-    assertPositiveTimelineNumber(scale, 'scale');
-    const clampedScale = this.clampZoomScale(scale);
-
-    this.state.zoomScale = clampedScale;
-    const clampedScroll = Math.max(0, Math.min(this.state.scrollLeft, this.maxScrollLeft));
-    this.state.scrollLeft = clampedScroll;
-
-    this.emit('zoom:change', clampedScale);
-    this.emitScrollChange();
-    this.emit('render');
-    this.emit('state:settled');
-  }
-
-  /**
-   * Sets horizontal scroll offset, clamped to the valid scroll range.
-   *
-   * @param scroll - Desired horizontal scroll offset in pixels.
-   */
-  setScrollLeft(scroll: number) {
-    assertNonNegativeTimelineNumber(scroll, 'scroll');
-    const clamped = Math.max(0, Math.min(scroll, this.maxScrollLeft));
-    this.state.scrollLeft = clamped;
-    this.emitScrollChange();
-    this.emit('render');
-    this.emit('state:settled');
-  }
-
-  /**
-   * Sets vertical scroll offset, clamped to the valid track stack scroll range.
-   *
-   * @param scroll - Desired vertical scroll offset in pixels.
-   */
-  setScrollTop(scroll: number) {
-    assertNonNegativeTimelineNumber(scroll, 'scroll');
-    this.state.scrollTop = scroll;
-    this.clampScrollTop();
-    this.emitScrollChange();
-    this.emit('render');
-    this.emit('state:settled');
-  }
-
-  /**
-   * Sets or clears an explicit timeline duration.
-   *
-   * When a duration is set, zoom, scroll, and playhead are clamped to that
-   * duration instead of the dynamic maximum clip end.
-   *
-   * @param duration - Explicit duration, or `undefined` to use clip content bounds.
-   */
-  setDuration(duration: RationalTime | undefined) {
-    if (duration !== undefined) {
-      assertValidRationalTime(duration, 'duration');
-    }
-    this.state.duration = duration;
-
-    // clamp playhead and scroll if duration changed
-    if (duration !== undefined) {
-      if (compareRational(this.state.playheadTime, duration) > 0) {
-        this.updatePlayhead(duration);
-      }
-      this.setZoomScale(this.state.zoomScale); // re-clamp zoom and scroll
-    } else if (this.hasZoomConstraints()) {
-      this.setZoomScale(this.state.zoomScale); // re-clamp zoom if just removing duration
-    } else {
-      this.setScrollLeft(this.state.scrollLeft); // re-clamp scroll if just removing duration
-    }
-
-    this.emit('render');
-    this.emit('state:settled');
-  }
-
-  /**
-   * Stores the visible timeline viewport width.
-   *
-   * @param width - Viewport width in pixels.
-   */
-  setViewportWidth(width: number) {
-    assertNonNegativeTimelineNumber(width, 'width');
-    this.state.viewportWidth = width;
-    if (this.state.duration !== undefined || this.hasZoomConstraints()) {
-      this.setZoomScale(this.state.zoomScale); // re-clamp zoom scale based on new width
-    }
-    this.emitViewportResize();
-    this.emit('state:settled');
-  }
-
-  /**
-   * Stores the visible timeline viewport height.
-   *
-   * @param height - Viewport height in pixels.
-   */
-  setViewportHeight(height: number) {
-    assertNonNegativeTimelineNumber(height, 'height');
-    this.state.viewportHeight = height;
-    const scrollChanged = this.clampScrollTop();
-    if (scrollChanged) {
-      this.emitScrollChange();
-    }
-    this.emitViewportResize();
-    this.emit('state:settled');
   }
 }
